requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.8
django-neomodel==0.0.7
djangorestframework==3.11.2
gunicorn==20.0.4
neo4j-driver==4.3.1
neomodel==4.0.4
pytz==2021.1
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.2
whitenoise==3.3.1
